Removed Final_Project_Client namespace in lieu of regular MonoBehavior inheritance. and UnityEngine namespace

+++ PoppinRobots.sql
@@ -1,4 +1,4 @@
-DROP DATABASE PoppinRobots;
+IF EXISTS DROP DATABASE PoppinRobots;
 CREATE DATABASE PoppinRobots;
 USE PoppinRobots;
 CREATE TABLE Cosmetics (
